+++ CHANGELOG.md
@@ -5,6 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## 0.16.0 - 2026-03-13
+### Added
+- display pending IRM
+- display IRM history
+
+## Updated
+- display verification page based on silo address
+
 ## 0.15.1 - 2026-03-10
 ### Updated
 - improve visualy verification page